Confirm Modal Pop Up created

+++ src/app/forms/components/section/section.component.ts
@@ -24,6 +24,8 @@ import {
 import { State } from 'src/app/forms/state/builder/builder-state.selectors';
 import { SectionEvent, Section } from 'src/app/interfaces';
 import { BuilderConfigurationActions } from '../../state/actions';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmModalPopupComponent } from 'src/app/components/race-dynamic-form/confirm-modal-popup/confirm-modal-popup/confirm-modal-popup.component';
 
 @Component({
   selector: 'app-section',
@@ -104,7 +106,11 @@ export class SectionComponent implements OnInit, OnDestroy {
   private _sectionQuestionsCount: number;
   private onDestroy$ = new Subject();
 
-  constructor(private fb: FormBuilder, private store: Store<State>) {}
+  constructor(
+    private fb: FormBuilder,
+    private dialog: MatDialog,
+    private store: Store<State>
+  ) {}
 
   ngOnInit() {
     this.sectionForm.valueChanges
@@ -157,11 +163,37 @@ export class SectionComponent implements OnInit, OnDestroy {
   }
 
   deleteSection() {
-    this.sectionEvent.emit({
-      pageIndex: this.pageIndex,
-      sectionIndex: this.sectionIndex,
-      section: this.sectionForm.getRawValue(),
-      type: 'delete'
+    if (!this.isTemplate) {
+      this.sectionEvent.emit({
+        pageIndex: this.pageIndex,
+        sectionIndex: this.sectionIndex,
+        section: this.sectionForm.getRawValue(),
+        type: 'delete'
+      });
+      return;
+    }
+    const dialogRef = this.dialog.open(ConfirmModalPopupComponent, {
+      maxWidth: 'max-content',
+      maxHeight: 'max-content',
+      data: {
+        popupTexts: {
+          primaryBtnText: 'yes',
+          secondaryBtnText: 'cancel',
+          title: 'confirmDelete?',
+          subtitle:
+            'deletingTheSectionWillUnlinkTheElementsCreatedFromThisTemplate'
+        }
+      }
+    });
+    dialogRef.afterClosed().subscribe((data) => {
+      if (data === 'primary') {
+        this.sectionEvent.emit({
+          pageIndex: this.pageIndex,
+          sectionIndex: this.sectionIndex,
+          section: this.sectionForm.getRawValue(),
+          type: 'delete'
+        });
+      }
     });
   }
